\name{data_preprocess}
\alias{data_preprocess}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
titolo
}
\description{
descrivere cosa fa
}
\usage{
data_preprocess(XT, XS, YT, miss.cutoff = 0.995, num.min.lvs = 7, rare.cutoff = ceiling(nrow(XT) * 0.01), cat.trans = "None", num.trans = "None", num.miss = "None")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{XT}{
 \code{XT} 
}
  \item{XS}{
 \code{XS} 
}
  \item{YT}{
Describe \code{YT} here
}
  \item{miss.cutoff}{
Describe \code{miss.cutoff} here
}
  \item{num.min.lvs}{
Describe \code{num.min.lvs} here
}
  \item{rare.cutoff}{
Describe \code{rare.cutoff} here
}
  \item{cat.trans}{
Describe \code{cat.trans} here
}
  \item{num.trans}{
Describe \code{num.trans} here
}
  \item{num.miss}{
Describe \code{num.miss} here
}
}
\details{
 If necessary, more details than the description above 
}
\value{cosa restituisce la funzione
}
\references{
se ci sono
}
\author{
dario.solari@gmail.com
}
\note{
note qui
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{help}}
}
\examples{
## esempi qui
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
